% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fn_test.R
\name{Fn_test}
\alias{Fn_test}
\title{Determine the number of non-Gaussian and Gaussian factors in HMCFA}
\usage{
Fn_test(
  X,
  scale = F,
  rmax = 8,
  L = 100,
  method = c("ER", "GER3", "GER4", "JJR"),
  ...
)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{scale}{logical. If \code{TRUE}, the variance of columns of X are normalized to 1 before Factor Number test.}

\item{L}{An integer. Maximum iteration for JJR approach, default to 100}

\item{method}{Method to use: "\code{ER}", Ahn and Horenstein(2013)'s ER and GR estimators; "\code{GER3}", \code{GER3} and
\code{GGR3} estimators; "\code{GER3}", \code{GER4} and \code{GGR4} estimators; "\code{JJR}", Jondeau et al.(2018)'s threshold approach}

\item{...}{Any other passthru parameters}
}
\value{
The number of non-Gaussian and Gaussian factors determined by selected approach.
}
\description{
Determine the number of non-Gaussian and Gaussian factors
}
\examples{
n = 20;t = 100;k = 2;msig_e = 7;
par_f = c(0.5,0,2,2,Inf,Inf);
par_e = c(1,0,2,Inf,0,0,0);
rho_ar = c(0.5,0.2);
data = SimHFA(n,t,k,par_f,par_e,rho_ar,msig_e)$X;
Fn_test(data,method = "GER3")
}
