% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOFA.als.R
\name{HOFA.als}
\alias{HOFA.als}
\title{Implementation for Alternating Least Square (ALS) algorithm in HOFA}
\usage{
HOFA.als(
  X,
  method = c("HFA3", "HFA4"),
  gamma,
  rh,
  rg,
  rmax = 8,
  eps = 10^-8,
  ...
)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{method}{Algorithm to use: "\code{HFA3}" to estimate factors by third-order multi-cumulant,
"\code{HFA4}" to estimate factors by fourth-order multi-cumulant.}

\item{gamma}{A weighted vector, a 2x1 vector for \code{HFA3} and a 3x1 vector for \code{HFA4}}

\item{rh}{The number of non-Gaussian factors}

\item{rg}{The number of Gaussian factors}

\item{rmax}{The maximum number of factors}

\item{eps}{The iteration error, default to 10^-8}

\item{...}{Any other parameters}
}
\value{
Estimated factors, factor loadings and commmon components.
}
\description{
Implementation for Alternating Least Square (ALS) algorithm in HOFA
}
\examples{
n = 100;t = 200;k = 2;
par_f = c(0.5,0,2,2,Inf,Inf);
par_e = c(1,0,2,Inf,0,0,0);
rho_ar = c(0.5,0.2);
data = HOFA.sim(n,t,k,par_f,par_e,rho_ar)$X;
HOFA.als(data,method = "HFA3", gamma = c(1,1),rh = 1, rg = 1)
}
