% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hofa.sim.R
\name{hofa.DGP1}
\alias{hofa.DGP1}
\title{Generate panel data from [H]igher-[O]rder multi-cumulant [F]actor [A]nalysis of DGP1}
\usage{
hofa.DGP1(n, t, k, par_f, par_e, alpha, rho_f, rho_e = 0.2, ...)
}
\arguments{
\item{n}{An integer, the number of variables.}

\item{t}{An integer, the number of observations.}

\item{k}{An integer, the number of factors.}

\item{par_f}{A list of four elements, the parameters of factor distribution, the first list is sig_f = (sig_f1,sig_f2,...), the second list is eta_f = (eta_f1,eta_f2,...), the third list is p_f = (p_f1,p_f2,...), the fourth list is q_f = (q_f1,q_f2,...).}

\item{par_e}{A list of four elements, the parameters of error distribution, par_e = list(sig_e,eta_e,p_e,q_e).}

\item{alpha}{The eigenvalue rate of error terms.}

\item{rho_f}{A k x 1 vector, the auto-regressive parameters of factors.}

\item{rho_e}{The auto-regressive parameters of errors.}

\item{...}{Any other parameters.}
}
\value{
A T x N matrix, generated by DGP1 of hofa.
}
\description{
Generate panel data from [H]igher-[O]rder multi-cumulant [F]actor [A]nalysis of DGP1
}
\examples{
n = 100
t = 200
k = 2
par_f = list(rep(1,k),rep(0.8,k),rep(1,k),rep(Inf,k))
par_e = list(1,0,2,Inf)
rho_f = c(0.5,0.2)
alpha = 0
hofa.DGP1(n,t,k,par_f,par_e,alpha,rho_f,rho_e = 0.2)
}
