% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-description.R
\docType{package}
\name{hofa}
\alias{hofa}
\title{hofa: Econometric tools for higher-order multi-cumulant factor analysis}
\description{
The hofa package provides two important functions:
\code{GER.sel} and \code{hofa.als}.
}
\section{hofa functions}{

\code{GER.sel} provides Generalized Eigenvalue Ratio test for selecting the number
of non-Gaussian factors in Higher-order multi-cumulant factor model. The
documentation of the \code{GER.sel} function can be loaded by \code{?GER.sel}.

\code{hofa.als} provides Alternating Least Square algorithm by giving the number of
non-Gaussian and Gaussian factors. The documentation of the `ALS` function
can be loaded by \code{?hofa.als}.
}

