% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M2.mle.R
\name{M2.mle}
\alias{M2.mle}
\title{Estimating the latent factors using maximum likelihood methods}
\usage{
M2.mle(
  X,
  r,
  scale = F,
  method = c("ML", "QML", "ML-GLS", "ML-ITE", "ML-EM"),
  eps = 10^-6,
  ar.order = 1,
  ...
)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{r}{The number of factors.}

\item{scale}{logical. If \code{TRUE}, the variance of columns of X are normalized to 1 before factor estimation.}

\item{method}{Method to use: "\code{ML}", Maximum Likelihood method in Bai and Li (2012); "\code{QML}", Quasi-Maximum Likelihood method in Bai and Li (2013); "\code{ML-GLS}", Bai and Li (2013)'s two step "ML-GLS" method;
"\code{ML-ITE}", Bai and Li (2013)'s Iterative "ML-GLS" method; "\code{ML-EM}", Bai and Li (2013)'s EM algorithm for Maximum Likelihood.}

\item{eps}{The iteration error, default to 10^-8. Available for Maximum Likelihood methods.}

\item{ar.order}{An integer. Auto regression lag for the idiosyncratic errors in \code{ML-GLS} and \code{ML-ITE}. If it is null, \code{ar.order} will be selected by AIC.}

\item{...}{Any other parameters.}
}
\value{
A list of factors, factor loadings and other information, see below.
\itemize{
  \item{\code{f}}{  Estimated factors.}
  \item{\code{u}}{  Estimated factor loadings.}
  \item{\code{e}}{  Estimated errors.}
  \item{\code{m2e}}{  Diagonal elements of the covariance matrix of errors, only provided in \code{PCA} and \code{ML}.}
  \item{\code{rho}}{  Auto regression coefficients of errors, only provided in \code{ML-GLS}, \code{ML-ITE} and \code{ML-EM}.}
}
}
\description{
Estimating the latent factors and factor loadings in high dimensional factor model using maximum likelihood methods based on the covariance or correlation matrix.
}
\examples{
n = 100
t = 200
k = 2
par_f = c(0.5,0,2,2,Inf,Inf)
par_e = c(1,0,2,Inf,0,0,0)
rho_ar = c(0.5,0.2)
data = hofa.sim(n,t,k,par_f,par_e,rho_ar)$X
M2.mle(data,r = 2,method = "ML-EM")
}
