% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M2.pca.R
\name{M2.pca}
\alias{M2.pca}
\title{Estimating the latent factors using principal component methods}
\usage{
M2.pca(
  X,
  C = NULL,
  r,
  center = F,
  scale = F,
  method = c("PCA", "P-PCA"),
  J = NULL,
  ...
)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{C}{Characteristics, a matrix with n rows (variables) and d columns (characteristics), used in Projected PCA.}

\item{r}{The number of factors.}

\item{center}{logical. If \code{TRUE}, the mean of columns of X are normalized to 0 before factor estimation.}

\item{scale}{logical. If \code{TRUE}, the variance of columns of X are normalized to 1 before factor estimation.}

\item{method}{Method to use: "\code{PCA}", Basic Principal Component Analysis; "\code{P-PCA}", Fan et al.(2016)'s Projected PCA.}

\item{J}{The number of sieve terms in Projected PCA. Default to use the criterion in Fan et al.(2016).}

\item{...}{Any other parameters.}
}
\value{
A list of factors, factor loadings and other information, see below.
\itemize{
  \item{\code{f}}{  Estimated factors.}
  \item{\code{u}}{  Estimated factor loadings.}
  \item{\code{e}}{  Estimated errors.}
  \item{\code{ev}}{ Eigenvalues of covariance matrix.}
  \item{\code{G}}{  Estimated non-parametric functions, only provided in \code{P-PCA}.}
  \item{\code{gamma}}{ Errors in factor loading matrix, only provided in \code{P-PCA}.}
}
}
\description{
Estimating the latent factors and factor loadings in high dimensional factor model using principal component methods based on the covariance or correlation matrix.
}
\examples{
n = 100;t = 10;d = 1;r = 3;
g1 = function(x){x^3-2*x}
g2 = function(x){x^2-1}
g3 = function(x){x}
C = matrix(rnorm(n*d),n,d);W = matrix(NA,n,r)
W[,1] <- g1(C);W[,2] <- g2(C);W[,3] <- g3(C)
FF = matrix(rnorm(t*r),t,r)
EE = matrix(rnorm(t*n),t,n)
X = W\%*\%t(FF) + t(EE)
M2.pca(t(X),C = C,r,method = "P-PCA",J = 4)
}
