% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M2.select.R
\name{M2.select}
\alias{M2.select}
\title{Determine the number of factors based on the covariance or correlation matrix}
\usage{
M2.select(
  X,
  scale = F,
  rmax = 8,
  method = c("ER", "GR", "BN-IC3", "BN-PC3", "BIC3", "ON", "ACT"),
  modified = F,
  ...
)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{scale}{logical. If \code{TRUE}, the variance of columns of X are normalized to 1 before factor number test.}

\item{rmax}{The maximum number of factors.}

\item{method}{Method to use: "\code{ER}" and "\code{GR}" are Ahn and Horenstein(2013)'s ER and GR estimators;
"\code{BN-IC3}","\code{BN-PC3}" and "\code{BIC3}" are Bai and Ng(2002)'s IC3, PC3 and BIC3 estimators, respectively;
"\code{ON}" is Onatski(2010)'s estimator; "\code{ACT}" is Fan et al.(2020)'s Adjusted Correlation Thresholding method.}

\item{modified}{logical. Only available for "\code{ER}" and "\code{GR}". If \code{TRUE}, we use modified estimators which can test zero factor.}

\item{...}{Any other parameters.}
}
\value{
The number of factors determined by selected approach and the eigenvalues of the covariance matrix.
}
\description{
Determine the number of factors based on the covariance or correlation matrix
}
\examples{
n = 100
t = 200
k = 2
par_f = list(rep(1,k),rep(0.8,k),rep(1,k),rep(Inf,k))
par_e = list(1,0,2,Inf)
rho_f = c(0.5,0.2)
par_cove = list(beta = 0.2,J = n/10,rho = 0.2,msig_e = c(1,5))
data = hofa.DGP2(n,t,k,par_f,par_e,par_cove,rho_f)$X
M2.select(data,method = "ER")
}
