% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3.als.R
\name{M3.als}
\alias{M3.als}
\title{Third-order Alternating Least Square (ALS) algorithm}
\usage{
M3.als(X, scale = FALSE, gamma = NULL, rh, rg, eps = 10^-8, ...)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{scale}{logical. If \code{TRUE}, the variance of columns of X are normalized to 1 before factor estimation.}

\item{gamma}{A weighted vector, default to (0,1).}

\item{rh}{The number of non-Gaussian factors.}

\item{rg}{The number of Gaussian factors.}

\item{eps}{The iteration error, default to 10^-8.}

\item{...}{Any other parameters.}
}
\value{
Estimated factors, factor loadings and errors.
}
\description{
Estimating the latent factors using Alternating Least Square (ALS) algorithm based on third-order multi-cumulant.
}
\examples{
n = 100
t = 200
k = 2
par_f = list(rep(1,k),rep(0.8,k),rep(1,k),rep(Inf,k))
par_e = list(1,0,2,Inf)
rho_f = c(0.5,0.2)
par_cove = list(beta = 0.2,J = n/10,rho = 0.2,msig_e = c(1,5))
data = hofa.DGP2(n,t,k,par_f,par_e,par_cove,rho_f)$X
M3.als(data,rh = 1, rg = 1)
}
