% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M4.gmm.R
\name{M4.gmm}
\alias{M4.gmm}
\title{Estimating the latent factors using fourth-order generalized moment methods}
\usage{
M4.gmm(
  X,
  r,
  kappa = 0,
  initial = c("PCA", "MLE"),
  skewness = FALSE,
  W_diag = FALSE,
  identity = FALSE,
  delta = NULL,
  eps = 10^-6,
  ...
)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{r}{The number of factors.}

\item{kappa}{An integer. The weight between \code{M} and \code{VWV}, \code{M} denotes the covariance matrix and \code{VWV} denotes the GMM matrix. \code{kappa} default to 0.}

\item{initial}{The method used to initialize the factor loadings and the moments of error. \code{PCA} denotes Principal Component Analysis and \code{MLE} denotes Maximum Likelihood Estimation in Bai and Li(2012).}

\item{skewness}{Logical. If \code{TRUE}, the skewness moment structure equations will be used.}

\item{W_diag}{Logical. If \code{TRUE}, the weight matrix \code{W} only has nonzero diagonal elements, default to \code{FALSE}.}

\item{identity}{Logical. If \code{TRUE}, the moments of errors are unified to identity, default to \code{FALSE}.}

\item{delta}{An integer. The hard threshold value of \code{W} matrix in order to guarantee non-singularity of \code{W}, default to 1/log(t).}

\item{eps}{The iteration error, default to 10^-6. Available for initializing the estimators by Maximum Likelihood method.}

\item{...}{Any other parameters.}
}
\value{
A list of factors, factor loadings and other information, see below.
\itemize{
  \item{\code{f}}{  Estimated factors.}
  \item{\code{u}}{  Estimated factor loadings.}
  \item{\code{e}}{  Estimated errors.}
  \item{\code{ev}}{ Eigenvalues of covariance matrix.}
}
}
\description{
Estimating the latent factors and factor loadings in high dimensional factor model using generalized moment methods based on the third-order cumulant.
}
\examples{
n = 100
t = 200
k = 2
par_f = c(0.5,0,2,2,Inf,Inf)
par_e = c(1,0,2,Inf,0,0,0)
rho_ar = c(0.5,0.2)
data = hofa.sim(n,t,k,par_f,par_e,rho_ar)$X
M4.gmm(data,r = 2,kappa = 0,initial = "PCA")
}
