% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M4.select.R
\name{M4.select}
\alias{M4.select}
\title{Determine the number of non-Gaussian and Gaussian factors in hofa}
\usage{
M4.select(
  X,
  scale = F,
  rmax = 8,
  method = c("GER4", "GGR4", "JJR4"),
  modified = F,
  L = 100,
  ...
)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{scale}{logical. If \code{TRUE}, the variance of columns of X are normalized to 1 before Factor Number test.}

\item{rmax}{The maximum number of factors.}

\item{method}{Method to use: "\code{GER4}" and "\code{GGR4}", Lu et al.(2021)'s GER4 and GGR4 estimators; "\code{JJR4}", Jondeau et al.(2018)'s threshold approach.}

\item{modified}{logical. Only available for "\code{GER4}" and "\code{GGR4}". If \code{TRUE}, we use modified estimators which can test zero factor.}

\item{L}{An integer. Maximum iteration for JJR4 approach, default to 100.}

\item{...}{Any other parameters.}
}
\value{
The number of non-Gaussian and Gaussian factors determined by selected approach.
}
\description{
Determine the number of non-Gaussian and Gaussian factors in hofa
}
\examples{
n = 100;t = 200;k = 2;
par_f = c(0.5,0,2,2,Inf,Inf);
par_e = c(1,0,2,Inf,0,0,0);
rho_ar = c(0.5,0.2);
data = hofa.sim(n,t,k,par_f,par_e,rho_ar)$X;
M4.select(data,method = "GER4")
}
