% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHOFA.R
\name{SimHOFA}
\alias{SimHOFA}
\title{Genereate panel data from [H]igher-[O]rder multi-cumulant [F]actor [M]odel (HOFA)}
\usage{
SimHOFA(n, t, k, par_f, par_e, rho_ar, msig_e, ...)
}
\arguments{
\item{n}{An integer, the number of variables}

\item{t}{An integer, the number of observations}

\item{k}{An integer, the number of factors}

\item{par_f}{A kx3 vector, the parameters of factor distribution, par_f = c(eta_f1,eta_f2,...,p_f1,p_f2,...,q_f1,q_f2,...)}

\item{par_e}{A vector, the parameters of error distribution. par_e = c(theta,eta_e,p_e,q_e,beta,J,rho)}

\item{rho_ar}{A kx1 vector, the auto-regressive parameters of factors}

\item{msig_e}{The maximum variance of error terms}

\item{...}{Any other parameters}
}
\value{
A T x N matrix, generated by Higher-order multi-cumulant Factor Model.
}
\description{
Genereate panel data from [H]igher-[O]rder multi-cumulant [F]actor [M]odel (HOFA)
}
\examples{
n = 100;t = 200;k = 2;msig_e = 7;
par_f = c(0.5,0,2,2,Inf,Inf);
par_e = c(1,0,2,Inf,0,0,0);
rho_ar = c(0.5,0.2);
SimHOFA(n,t,k,par_f,par_e,rho_ar,msig_e)
}
