% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-description.R
\docType{package}
\name{HOFA}
\alias{HOFA}
\title{HOFA: A package for implementing Higher-order multi-cumulant Factor Analysis}
\description{
The HOFA package provides two important functions:
GER.sel and ALS.
}
\section{HOFA functions}{

`GER.sel` provides Generalized Eigenvalue Ratio test for selecting the number
of non-Gaussian factors in Higher-order multi-cumulant factor model. The
documentation of the `GER.sel` function can be loaded by `?GER.sel`.

`ALS` provides Alternating Least Square algorithm by giving the number of
non-Gaussian and Gaussian factors. The documentation of the `ALS` function
can be loaded by `?ALS`.
}

