% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M4.als.R
\name{M4.als}
\alias{M4.als}
\title{Implementation for Alternating Least Square (ALS) algorithm based on fourth-order multi-cumulant.}
\usage{
M4.als(X, scale = FALSE, gamma = NULL, rh, rg, eps = 10^-8, ...)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{scale}{logical. If \code{TRUE}, the variance of columns of X are normalized to 1 before factor estimation.}

\item{gamma}{A weighted vector, defualt to (0,0,1).}

\item{rh}{The number of non-Gaussian factors.}

\item{rg}{The number of Gaussian factors.}

\item{eps}{The iteration error, default to 10^-8.}

\item{...}{Any other parameters.}
}
\value{
Estimated factors, factor loadings and errors.
}
\description{
Implementation for Alternating Least Square (ALS) algorithm based on fourth-order multi-cumulant.
}
\examples{
n = 100;t = 200;k = 2;
par_f = c(0.5,0,2,2,Inf,Inf);
par_e = c(1,0,2,Inf,0,0,0);
rho_ar = c(0.5,0.2);
data = hofa.sim(n,t,k,par_f,par_e,rho_ar)$X;
M4.als(data,rh = 1, rg = 1)
}
